+++ index.php
@@ -31,8 +31,14 @@ if (!function_exists('add_action')) {
 include('includes/activate.php');
 include('includes/init.php');
 include('includes/admin/init.php');
+include('process/save-term.php');
+include( 'includes/admin/forms.php' );
 
 // Hooks
 register_activation_hook(__FILE__, 'nsm_activate_plugin');
 add_action('init', 'nsm_init', 0);
 add_action('admin_init', 'nsm_admin_init');
+add_action('saved_municipality', 'nsm_save_term_admin', 10, 3);
+add_action('municipality_add_form_fields', 'nsm_add_form_fields_new', 10, 1);
+add_action('municipality_edit_form_fields', 'nsm_add_form_fields_edit', 10, 2);
+add_action('admin_footer', 'nsm_javascript');