+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Upcoming [1.2.0]
+
+### Added
+
+- Meta data for municipalities, including `long_title` and `partners`
+
 ## [1.1.2] - 2020-04-29
 
 ### Fixed
@@ -30,4 +36,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- `municipality` taxanomy
+- `municipality` taxonomy